Add a way to run tests to generate config values
Write module 'configure' (configure.erl) and place it in the same
directory with rebar.config.

In rebar config, specify values of form
  {test, Fun, Args}
  Fun = atom()
  Args = list()

When rebar is run, the forms will be replaced with values returned
by erlang:apply(configure, Fun, Args).

Values will be cached in rebar.cache and the cache is refreshed when
rebar.config or configure.erl modification timestamps are newer
than rebar.cache.